add impressions, clicks, visits statis

+++ Service/tracking/domain.go
@@ -10,9 +10,9 @@ import (
 
 // ReferrerDomainStatisKey AdReferrerDomainStatis表里面的Unique Key部分
 type ReferrerDomainStatisKey struct {
-	UserID         int
-	Timestamp      int
-	CampaignID     int
+	UserID         int64
+	Timestamp      int64
+	CampaignID     int64
 	ReferrerDomain string
 }
